fix: getAuthenticatedUser を src/lib/apiAuth.ts に切り出し共有化
