+++ src/lib/apiAuth.ts
@@ -1,0 +1,15 @@
+import { type NextRequest } from "next/server";
+
+import { prisma } from "@/lib/prisma";
+
+export async function getAuthenticatedUser(req: NextRequest) {
+  const authHeader = req.headers.get("authorization");
+  if (!authHeader?.startsWith("Bearer ")) return null;
+  const apiKey = authHeader.slice(7).trim();
+  if (!apiKey) return null;
+  const user = await prisma.user.findUnique({
+    where: { apiKey },
+    select: { id: true, role: true, isActive: true },
+  });
+  return user?.isActive ? user : null;
+}
